+++ cross_engine_overhead/dependent/add_rmsnorm.py
@@ -12,7 +12,7 @@ M = 4096
 class AddOp(nn.Module):
     def __init__(self, H):
         super().__init__()
-        self.register_buffer('bias', torch.randn(1, H) * 0.01)
+        self.register_buffer('bias', torch.randn(1, H, dtype=torch.bfloat16) * 0.01)
     def forward(self, x):
         return x + self.bias
 
@@ -20,7 +20,7 @@ class AddOp(nn.Module):
 class RMSNormOp(nn.Module):
     def __init__(self, H):
         super().__init__()
-        self.register_buffer('weight', torch.ones(H))
+        self.register_buffer('weight', torch.ones(H, dtype=torch.bfloat16))
         self.eps = 1e-6
     def forward(self, x):
         rms = torch.sqrt(torch.mean(x ** 2, dim=-1, keepdim=True) + self.eps)
@@ -38,7 +38,7 @@ class DependentPair(nn.Module):
 
 def main():
     model = DependentPair().eval()
-    x = torch.randn(M, H)
+    x = torch.randn(M, H, dtype=torch.bfloat16)
     workdir = f"/tmp/neuron_cross_add_rmsnorm_{M}x{H}"
     print(f"Compiling add→rmsnorm ({M},{H})...", end=" ", flush=True)
     torch_neuronx.trace(model, (x,), compiler_workdir=workdir)
