+++ cross_engine_overhead/dependent/silu_mul.py
@@ -17,7 +17,7 @@ class SiLUOp(nn.Module):
 class MulOp(nn.Module):
     def __init__(self, H):
         super().__init__()
-        self.register_buffer('gate', torch.randn(1, H) * 0.01)
+        self.register_buffer('gate', torch.randn(1, H, dtype=torch.bfloat16) * 0.01)
     def forward(self, x):
         return x * self.gate
 
@@ -33,7 +33,7 @@ class DependentPair(nn.Module):
 
 def main():
     model = DependentPair().eval()
-    x = torch.randn(M, H)
+    x = torch.randn(M, H, dtype=torch.bfloat16)
     workdir = f"/tmp/neuron_cross_silu_mul_{M}x{H}"
     print(f"Compiling silu→mul ({M},{H})...", end=" ", flush=True)
     torch_neuronx.trace(model, (x,), compiler_workdir=workdir)
