+++ cross_engine_overhead/independent/silu_mul.py
@@ -16,7 +16,7 @@ class SiLUOp(nn.Module):
 class MulOp(nn.Module):
     def __init__(self, H):
         super().__init__()
-        self.register_buffer('gate', torch.randn(1, H) * 0.01)
+        self.register_buffer('gate', torch.randn(1, H, dtype=torch.bfloat16) * 0.01)
     def forward(self, x):
         return x * self.gate
 
@@ -32,8 +32,8 @@ class IndependentPair(nn.Module):
 
 def main():
     model = IndependentPair().eval()
-    x1 = torch.randn(M, H)
-    x2 = torch.randn(M, H)
+    x1 = torch.randn(M, H, dtype=torch.bfloat16)
+    x2 = torch.randn(M, H, dtype=torch.bfloat16)
     workdir = f"/tmp/neuron_indep_silu_mul_{M}x{H}"
     print(f"Compiling silu_mul independent ({M},{H})...", end=" ", flush=True)
     torch_neuronx.trace(model, (x1, x2), compiler_workdir=workdir)
